requirements.txt
accelerate==1.7.0
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
bidict==0.23.1
bitsandbytes==0.45.5
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cryptography==45.0.2
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
eventlet==0.40.0
faiss-cpu==1.11.0
fastapi==0.115.12
filelock==3.18.0
Flask>=2.3.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
flask-pymongo==2.3.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fsspec==2025.5.0
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.73.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.61
langchain-text-splitters==0.3.8
langsmith==0.3.42
livereload==2.7.1
lxml==5.4.0
MarkupSafe==3.0.2
motor==3.7.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
oauthlib==3.2.2
orjson==3.10.18
packaging==24.2
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.13.0
PyMuPDF==1.26.1
PyMySQL==1.1.1
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.1
python-docx==1.1.2
python-dotenv==1.1.0
python-engineio==4.12.2
python-http-client==3.3.7
python-socketio==5.13.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sendgrid==6.12.2
sentence-transformers==4.1.0
setuptools==80.8.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.2
typing-inspection==0.4.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
werkzeug>=3.0.0
wsproto==1.2.0
WTForms==3.2.1
zstandard==0.23.0
